Yacarof/feat/tdp 5575 make multitenant users for it configurable (#1320)
(cherry picked from commit 416398d)
