+++ dataprep-test-api/src/test/java/org/talend/dataprep/qa/config/User.java
@@ -1,0 +1,26 @@
+package org.talend.dataprep.qa.config;
+
+public class User {
+
+    private String login;
+
+    private String password;
+
+    public String getLogin() {
+        return login;
+    }
+
+    public User setLogin(String login) {
+        this.login = login;
+        return this;
+    }
+
+    public String getPassword() {
+        return password;
+    }
+
+    public User setPassword(String password) {
+        this.password = password;
+        return this;
+    }
+}
