+++ dataprep-test-api/src/test/resources/application.properties
@@ -3,3 +3,4 @@ backend.api.url=http://localhost:8888
 
 # How long should we wait, in seconds, for asynchronous API service
 async.service.timeout = 10
+restassured.debug=false
